knowledge-graph-hackathon/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
airbyte==0.1.0
notion-pypi==1.0.0
spacy==3.7.2
neo4j==5.28.1
python-dotenv==1.1.0
pandas==2.1.3
httpx==0.25.0
python-multipart==0.0.6
